+++ src/rules/no-restricted-classes.ts
@@ -1,12 +1,5 @@
 import { defineRule } from '@oxlint/plugins'
-import {
-  extractFromJSXAttribute,
-  extractFromCallExpression,
-  extractFromTaggedTemplate,
-  extractFromVariableDeclarator,
-  DEFAULT_EXTRACTOR_CONFIG,
-  type ClassLocation,
-} from '../utils/extractors'
+import { createExtractorVisitors, type ClassLocation } from '../utils/extractors'
 import { splitClasses } from '../utils/class-splitter'
 import { safeOptions } from '../types'
 
@@ -114,19 +107,6 @@ export const noRestrictedClasses = defineRule({
       }
     }
 
-    return {
-      JSXAttribute(node) {
-        check(extractFromJSXAttribute(node, DEFAULT_EXTRACTOR_CONFIG))
-      },
-      CallExpression(node) {
-        check(extractFromCallExpression(node, DEFAULT_EXTRACTOR_CONFIG))
-      },
-      TaggedTemplateExpression(node) {
-        check(extractFromTaggedTemplate(node, DEFAULT_EXTRACTOR_CONFIG))
-      },
-      VariableDeclarator(node) {
-        check(extractFromVariableDeclarator(node, DEFAULT_EXTRACTOR_CONFIG))
-      },
-    }
+    return createExtractorVisitors(context, check)
   },
 })